requirements.txt
absl-py==0.10.0
APScheduler==3.6.3
astunparse==1.6.3
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
cycler==0.10.0
dill==0.3.2
dm-tree==0.1.5
fbm==0.3.0
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.1
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.2.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.3
Pillow==8.0.1
promise==2.3
protobuf==3.13.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-datasets==4.0.1
tensorflow-estimator==2.3.0
tensorflow-metadata==0.24.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.6.0
tornado==6.1
tqdm==4.51.0
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
